refactor AbstractFindMultipleByKeyOperation hierarchy
